+++ exasol_script_languages_container_tool/lib/models/language_activation_builder.py
@@ -1,6 +1,6 @@
 from typing import Dict, List
 
-from exasol_script_languages_container_tool.lib.tasks.upload.language_activation import (
+from exasol_script_languages_container_tool.lib.models.language_activation import (
     BuiltInLanguageDefinitionURL,
     LanguageDefinitionComponents,
     SLCLanguage,
